Moved range_image into common.
There were only two files - it didn't seem like enough for a
separate library. These are direct subclasses of pcl::PointCloud
so it seems to make sense to have them in the same library.

Also, now to add IO functions to the io/ library
we do not have to introduce a new dependency to io (it already
depends on common).



git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4393 a9d63959-f2ad-4865-b262-bf0e56cfafb6